% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidel_entropy.R
\name{fidel_entropy}
\alias{fidel_entropy}
\title{Calculate the entropy for each observation}
\usage{
fidel_entropy(X, probs, outCol = "entropy", scale = FALSE)
}
\arguments{
\item{X}{a data frame containing observations (i.e., instances, examples) on the rows, along with probability predictions for each
class.  May contain other columns as well.}

\item{probs}{A character vector, logical vector, or vector of numeric indexes that identify the columns in \code{X} that contain
the predicted probabilities of the classes.  This is passed to
\href{http://pnnl.github.io/docs-Smisc/rd.html#selectelements}{selectElements()} from the
\href{http://pnnl.github.io/docs-Smisc}{Smisc} package.}

\item{outCol}{A character string that indicates the name of the column that will contain the entropy.}

\item{scale}{A logical indicating whether the entropy should be linearly scaled.  See Details.}
}
\value{
The data frame \code{X} is returned, with an appended column that contains the entropy.
}
\description{
Calculate the entropy for each observation
}
\details{
The entropy is \emph{- sum_k (P_k * log(P_k)} where \emph{P_k} is the predicted probability that the true class is \emph{k}.

Note that for \code{scale = FALSE}, smaller is better, whereas for \code{scale = TRUE}, larger is better.

Setting \code{scale = TRUE} scales the entropy so that it is negative if the classifies does worse than guessing (i.e.,
worse than assinging uniform probabilities). Predicted probabilities with a 
minimum entropy (maximum information) have a scaled entropy of \code{1} (which occurs when one class
has a predicted probability of 1 and the remainder are 0). 

The linear transformation
that takes place when \code{scale = TRUE} maps the entropy, which ranges in \emph{[0,  log(K)]},
(\emph{K} being the number of classes), to \emph{0, 1]}, where \emph{0} indicates uniform probability assignments and \emph{1}
indicates all the probability was assigned to a single class.
Additional information about the scaling is available in Section 3.2.1 of Holmes et al. (2013).
}
\examples{
# Construct probability vectors that sum (piecewise) to 1
p1 <- runif(9)
p2 <- runif(9, 0, 1 - p1)
p3 <- 1 - p1 - p2

X <- data.frame(prob_7 = p1,
                prob_9 = p2,
                prob_2 = p3,
                extra = rnorm(9))

fidel_entropy(X, 1:3)
}
\references{
Holmes AE, Sego LH, Webb-Robertson BJ, et al. (2013). An Approach for Assessing the Signature Quality of
Various Chemical Assays when Predicting the Culture Media Used to Grow Microorganisms.  Pacific Northwest National Laboratory,
PNNL-22126.
}

