% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelFunctions.R
\name{fidelFunctions}
\alias{fidelFunctions}
\title{Lists the available functions that measure classifier fidelity}
\usage{
fidelFunctions(fidelFun = NULL)
}
\arguments{
\item{saUtilFun}{If not \code{NULL}, this should be a character string with the full name of the fidelity function,
or the abbreviated name that describes the function.  See Examples.}
}
\value{
By default, a vector with all available fidelity functions, or if a single fidelity function
is provided to \code{sa}, the name of the function is returned if it exists.  Otherwise an error is thrown.
}
\description{
By default, this function returns a vector with all of the available fidelity
functions that are available in the \code{maua} package. It can also be used to
verify that a fidelity function is available in the package.
}
\details{
All the single attribute
utility functions are of the form \emph{fidel_name} where \emph{name} is description of the function. Either
the full function name or the abbreviated, description, \emph{name}, can be supplied to the \code{fidelFun} argument.

\code{fidelFunctions()} facilitates the addition of new single attribute utility functions to the package.
}
\examples{
fidelFunctions()

# The single attribute exponential utility function.
# Note how both the full name of the function, "fidel_exp" and
# its abbreviation both work:
fidelFunctions("fidel_exp")
fidelFunctions("exp")
}
\seealso{
\code{\link{saUtilFunctions}}
}

