% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maua-package.R
\docType{package}
\name{maua-package}
\alias{maua}
\alias{maua-package}
\title{maua}
\description{
maua: Multi-Attribute Utility Assessment
}
\details{
\tabular{ll}{
Package: \tab maua\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2016-04-15\cr
License: \tab file LICENSE\cr
}
}
\examples{
help(package = maua)
}
\author{
Landon Sego

Maintainer: Landon Sego <Landon.Sego@pnnl.gov>
}
\keyword{package}

