% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saUtil_exp.R
\name{saUtil_exp}
\alias{saUtil_exp}
\title{The single attribute exponential utility function}
\usage{
saUtil_exp(z, theta = 0, zrange = range(z), urange = c(0, 1),
  certEquiv = NULL)
}
\arguments{
\item{z}{A numeric vector of attribute measurements, or values, in their native scale}

\item{theta}{A parameter that controls the convexity of the exponential utility function. For \code{theta > 0}
the utility is convex, for \code{theta < 0}, concave, and for \code{theta = 0}, it's linear.}

\item{zrange}{A vector with 2 elements containing the minimum and maximum possible values of the attribute.}

\item{urange}{A vector with 2 elements indicating the range of the utility function. \code{urange[1]}
is the mapping of \code{zrange[1]}, and \code{urange[2]} is the mapping of \code{zrange[2]}.}

\item{certEquiv}{A numerical value indicating the certainty equivalent, which, if provided, will
be used to solve for the value of \code{theta}--and that solution will overide the value of \code{theta}
in the function call. See Section 3.5.1.1 of the reference for more discussion on the certainty equivalent.}
}
\value{
An object of class \code{saUtilCall}, which includes the utility values of \code{z}. This object can be
printed or plotted via \code{\link{print.saUtilCall}} or \code{\link{plot.saUtilCall}}.
}
\description{
The single attribute exponential utility function
}
\details{
The exponential utility function may be increasing or decreasing (depending on the choice of
\code{urange}), and concave, convex, or linear (depending on the choice of \code{theta}).
}
\examples{
# Some "data"
x <- 1:10
# Increasing utility
saUtil_exp(x, theta = 3)

# Decreasing utility
u <- saUtil_exp(x, theta = 3, urange = c(1, 0))
print(u)
plot(u)
attributes(u)

# Obtain the value of theta using the certainty equivalent
theta0 <- attributes(saUtil_exp(x, certEquiv = 4))$parms$theta
print(theta0)
}
\references{
Holmes AE, Sego LH, Webb-Robertson BM, et al.  "An Approach for
Assessing the Signature Quality of Various Chemical Assays when
Predicting the Culture Media Used to Grow Microorganisms."  Technical
Report, Pacific Northwest National Laboratory, PNNL-22126.  February, 2013.
}

