% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saUtil_identity.R
\name{saUtil_identity}
\alias{saUtil_identity}
\title{The single attribute utility function that returns its argument}
\usage{
saUtil_identity(z)
}
\arguments{
\item{z}{A numeric vector of attribute measurements, or values, in their native scale}
}
\value{
An object of class \code{saUtilCall}, which is equal to \code{z}. This object can be
printed or plotted via \code{\link{print.saUtilCall}} or \code{\link{plot.saUtilCall}}.
}
\description{
The single attribute utility function that returns its argument
}
\examples{
# Some "data"
x <- 1:10
# Identity utility
saUtil_identity(x)
}

