% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saUtil_linear.R
\name{saUtil_linear}
\alias{saUtil_linear}
\title{The single attribute linear utility function}
\usage{
saUtil_linear(z, zrange = range(z), urange = c(0, 1))
}
\arguments{
\item{z}{A numeric vector of attribute measurements, or values, in their native scale}

\item{zrange}{A vector with 2 elements containing the minimum and maximum possible values of the attribute.}

\item{urange}{A vector with 2 elements indicating the range of the utility function. \code{urange[1]}
is the mapping of \code{zrange[1]}, and \code{urange[2]} is the mapping of \code{zrange[2]}.}
}
\value{
An object of class \code{saUtilCall}, which includes the utility values of \code{z}. This object can be
printed or plotted via \code{\link{print.saUtilCall}} or \code{\link{plot.saUtilCall}}.
}
\description{
The single attribute linear utility function
}
\details{
The linear function may be increasing or decreasing (depending on the choice of
\code{urange}).
}
\examples{
# Some "data"
x <- 1:10
# Increasing utility
saUtil_linear(x)

# Decreasing utility
u <- saUtil_linear(x, urange = c(1, 0))
print(u)
plot(u)
attributes(u)
}

