% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saUtilFunctions.R
\name{saUtilFunctions}
\alias{saUtilFunctions}
\title{Lists the available single attribute utility functions}
\usage{
saUtilFunctions(saUtilFun = NULL)
}
\arguments{
\item{method}{If supplied, a character string with the name of the single attribute utility function.}
}
\value{
By default, a vector with all available single attribute utility functions, or if a single attribute utility function
is provided to \code{saUtilFun}, \code{TRUE} or \code{FALSE} is returned depending on whether the function exists.
}
\description{
By default, this function returns a vector with all of the available single attribute
utility functions that are available in the \code{maua} package. It can also be used to
verify that a single attribute utility function is available in the package.
}
\details{
\code{saUtilFunctions()} facilitates the addition of new single attribute utility functions to the package.
}
\examples{
saUtilFunctions()

# The 'expUtil' function exists
saUtilFunctions("expUtil")

# But the 'noUtil'function does not
saUtilFunctions("noUtil")
}

